pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["LuybovB <grandprinceess7777@gmail.com>"]
readme = "README.md"
packages = [
    { include = "task_manager" },
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.13"
django-admin = "^2.0.2"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
dj-database-url = "^2.1.0"
django-bootstrap5 = "^23.3"
psycopg2-binary = "^2.9.7"
whitenoise = "^5.2.0"
django-filter = "^24.2"
django-widget-tweaks = "^1.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
